Fixing the SocialLinks icon visibility in light/dark mode
