+++ utils/context.ts
@@ -1,0 +1,3 @@
+import React from "react";
+
+export const DarkModeContext = React.createContext({ isDarkMode: false, toggle: () => {}})
